The jstats statistics have been moved to a separate project djutils-stats, as their use is independent of simulation and much more generic than dsol itself. As a result, the dsol statistics now extend the djutils-stats statistics. The nl.tudelft.simulation.jstats.statistics package has been removed from dsol-core. The DSOL statistics have been updated with the more advanced functions of the djutils-stats  project, such as the possibility to show percentiles, skewness and kurtosis. All swing-related classes have been moved from dsol-core to dsol-swing. Also the generic 'Swingable' class is now in dsol-swing. The table representations of dsol-core statistics have stayed, however, in dsol-core because the table representation itself is independent of its rendering (which could be using swing, on the web, or in a file). All demos have been updated to use the new statistics classes.

+++ trunk/dsol-core/src/main/java/nl/tudelft/simulation/dsol/formalisms/flow/Seize.java
@@ -159,7 +159,7 @@ public class Seize<A extends Comparable<A> & Serializable, R extends Number & Co
                 }
                 this.fireTimedEvent(Seize.QUEUE_LENGTH_EVENT, (double) this.queue.size(), this.simulator.getSimulatorTime());
                 R delay = this.simulator.getSimTime().diff(request.getCreationTime());
-                this.fireTimedEvent(Seize.DELAY_TIME, delay, this.simulator.getSimulatorTime());
+                this.fireTimedEvent(Seize.DELAY_TIME, delay.doubleValue(), this.simulator.getSimulatorTime());
                 this.releaseObject(request.getEntity());
                 return;
             }
